+++ aquadoggo/src/graphql/scalars/document_view_id_scalar.rs
@@ -5,7 +5,8 @@ use std::{fmt::Display, str::FromStr};
 use dynamic_graphql::{Error, Result, Scalar, ScalarValue, Value};
 use p2panda_rs::document::DocumentViewId;
 
-/// Document view id as a GraphQL scalar.
+/// The document view id of a p2panda document. Refers to a specific point in a documents history
+/// and can be used to deterministically reconstruct it's state at that time.
 #[derive(Scalar, Clone, Debug, Eq, PartialEq)]
 #[graphql(name = "DocumentViewId")]
 pub struct DocumentViewIdScalar(DocumentViewId);